setting up react router
