+++ src/frontend/package.json
@@ -14,7 +14,8 @@
     "@emotion/styled": "^11.11.0",
     "@mui/material": "^5.14.5",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
