+++ src/frontend/src/router/router.jsx
@@ -1,0 +1,12 @@
+import { Routes, Route } from 'react-router-dom';
+import Home from '../views/Home'
+import Dashboard from '../views/Dashboard'
+
+export default function ViewRouter() {
+    return (
+      <Routes>
+        <Route path="/" Component={Home} />
+        <Route path="/Dashboard" Component={Dashboard} />
+      </Routes>
+    );
+  }
