+++ src/frontend/src/views/Dashboard.jsx
@@ -1,0 +1,22 @@
+
+import '../App.css'
+
+function Dashboard () {
+
+
+
+
+  return (
+    <>
+      <div>
+       
+
+      <h2>SHR Dashboard</h2>
+
+      </div>
+
+    </>
+  )
+}
+
+export default Dashboard
